Authentication solution

+++ Planteer/plants/views.py
@@ -3,22 +3,32 @@ from django.http import HttpRequest, HttpResponse
 from .models import Plant, Comment, Country
 from .forms import PlantForm
 
+from django.core.paginator import Paginator
+from django.contrib import messages
+
 # Create your views here.
 
 
 def add_view(request:HttpRequest):
-    plant_form = PlantForm()
+        
+    if not request.user.is_staff:
+        messages.success(request, "only staff can add game", "alert-warning")
+        return redirect("main:home_view")
 
-    countries = Country.objects.all().order_by('name')
+    try:
+        plant_form = PlantForm()
+        countries = Country.objects.all().order_by('name')
+        if request.method == "POST":
+            plant_form = PlantForm(request.POST, request.FILES)
+            if plant_form.is_valid():
+                plant_form.save()
+                messages.success(request, "Plant added successfully", "alert-success")
+                return redirect('main:home_view')
+            return render(request, 'plants/add_plant.html', {'form':plant_form})
+    except Exception as e:
+        print(e)
+        messages.error(request, "Can't add plant right now, please try again later", "alert-danger")
 
-
-    if request.method == "POST":
-        plant_form = PlantForm(request.POST, request.FILES)
-        if plant_form.is_valid():
-            plant_form.save()
-            return redirect('main:home_view')
-        return render(request, 'plants/add_plant.html', {'form':plant_form})
-    
     return render(request, 'plants/add_plant.html', {"countries":countries})
 
 def plant_detail_view(request:HttpRequest, plant_id:int):
@@ -41,19 +51,26 @@ def plant_detail_view(request:HttpRequest, plant_id:int):
     return render(request, 'plants/plant_details.html', context)
 
 def update_view(request: HttpRequest, plant_id: int):
-    plant = Plant.objects.get(pk=plant_id)
-    countries = Country.objects.all().order_by('name')
 
+    if not request.user.is_staff:
+        messages.warning(request, "only staff can add game", "alert-warning")
+        return redirect("main:home_view")
 
-    if request.method == "POST":
-        plant_form = PlantForm(request.POST, request.FILES, instance=plant)
-        if plant_form.is_valid():
-            plant_instance = plant_form.save()
-
-            selected_countries = request.POST.getlist('countries')
-            plant_instance.countries.set(selected_countries)
-            return redirect('plants:plant_detail_view', plant_id=plant.id)
-        return render(request, 'plants/update_plant.html', {'form': plant_form, 'plant': plant, 'countries': countries})
+    try:
+        plant = Plant.objects.get(pk=plant_id)
+        countries = Country.objects.all().order_by('name')
+        if request.method == "POST":
+            plant_form = PlantForm(request.POST, request.FILES, instance=plant)
+            if plant_form.is_valid():
+                plant_instance = plant_form.save()
+                selected_countries = request.POST.getlist('countries')
+                plant_instance.countries.set(selected_countries)
+                messages.success(request, "Plant updated successfully", "alert-success")
+                return redirect('plants:plant_detail_view', plant_id=plant.id)
+            return render(request, 'plants/update_plant.html', {'form': plant_form, 'plant': plant, 'countries': countries})
+    except Exception as e:
+        print(e)
+        messages.error(request, "Can't update plant right now, please try again later", "alert-danger")
     
     plant_form = PlantForm(instance=plant)
     return render(request, 'plants/update_plant.html', {'form': plant_form, 'plant': plant, 'countries': countries})
@@ -99,22 +116,44 @@ def all_view(request:HttpRequest):
     if country_filter:
         plants = plants.filter(countries__id=country_filter)
 
+    page_number = request.GET.get("page", 1)
+    paginator = Paginator(plants,6)
+    plants_page = paginator.get_page(page_number)
 
-    return render(request, 'plants/all_plants.html', {'plants':plants, 'countries':countries, 'country_filter':country_filter })
+
+    return render(request, 'plants/all_plants.html', {'plants':plants_page, 'countries':countries, 'country_filter':country_filter })
 
 def delete_view(request:HttpRequest, plant_id):
-    plant = Plant.objects.get(pk=plant_id)
-    plant.delete()
+
+    if not request.user.is_staff:
+        messages.warning(request, "only staff can add game", "alert-warning")
+        return redirect("main:home_view")
+    
+    try:
+        plant = Plant.objects.get(pk=plant_id)
+        plant.delete()
+        messages.success(request, "Plant deleted successfully", "alert-success")
+    except Exception as e:
+        print(e)
+        messages.error(request, "Can't delete plant right now, please try again later", "alert-danger")
 
     return redirect("main:home_view")
 
 def add_comment_view(request:HttpRequest, plant_id):
+    if not request.user.is_authenticated:
+        messages.error(request, "Only registered user can add review","alert-danger")
+        return redirect("accounts:sign_in")
 
     if request.method == "POST":
         plant_object = Plant.objects.get(pk=plant_id)
-        new_comment= Comment(plant=plant_object, name = request.POST["name"], comment = request.POST["comment"])
+        new_comment= Comment(plant=plant_object, user = request.user, comment = request.POST["comment"])
         new_comment.save()
-
+        messages.success(request, "Comment added successfully", "alert-success")
     return redirect("plants:plant_detail_view", plant_id=plant_id)
 
+def plants_by_country(request:HttpRequest, country_id:int):
 
+    country = Country.objects.get(pk=country_id)
+    plants = country.plant_set.all()
+
+    return render(request, 'plants/plants_by_country.html', {'country':country, 'plants':plants})
